Use `NonZero` for type safety (#2180)

+++ crates/module-system/sov-test-utils/src/interface/inputs.rs
@@ -198,7 +198,7 @@ pub enum SequencerInfo {
     /// This is a preferred sequencer
     Preferred {
         /// The number of visible slots to advance
-        slots_to_advance: u64,
+        slots_to_advance: u8,
         /// The sequence number for this batch
         sequence_number: u64,
     },
